finish the project without the fungerPriny

+++ src/components/TravelsToUser.js
@@ -41,7 +41,7 @@ const TravelsToUser = (props) => {
     }
 
     const formatDate = (string) => {
-        var options = { year: 'numeric', month: 'long', day: 'numeric' };
+        var options = { year: 'numeric', month: 'numeric', day: 'numeric' };
         return new Date(string).toLocaleDateString([], options);
     }
